notificationAgent/requirements.txt
# requirements.txt (ADK Web用 最終版)
google-adk
deprecated
python-dotenv